This is where we landed at the end of the hackathon

+++ test.php
@@ -6,7 +6,7 @@ set_time_limit(0);
 date_default_timezone_set('America/Toronto');
 
 $datafolder = '../citybudget2012/';
-$startwithfilesthatbeginwith = 'a';
+$startwithfilesthatbeginwith = 't';
 
 /** Include path **/
 set_include_path(get_include_path() . PATH_SEPARATOR . './Classes/');
@@ -69,10 +69,25 @@ $revenues['sundryrevrow'] = "Sundry Revenues";
 $outputrevenuesfilename = "revenues.csv";
 $outputexpensesfilename = "expenses.csv";
 $outputpositionsfilename = "positions.csv";
-unlink($outputrevenuesfilename);
-unlink($outputexpensesfilename);
-unlink($outputpositionsfilename);
+$outputchartfilename = "chart.csv";
+$outputjsonfilename = "budget2012.json";
 
+$outputaslist = 1docs.google.com/feeds/download/spreadsheets/Export?key<FILE_ID>&exportFormat=csv&gid=0;
+$outputaschart = 0;
+$outputasjson = 0;
+
+if ($outputaslist) {
+	unlink($outputrevenuesfilename);
+	unlink($outputexpensesfilename);
+	unlink($outputpositionsfilename);
+} 
+if ($outputaschart) {
+	unlink($outputchartfilename);
+}
+if ($outputasjson) {
+	unlink($outputjsonfilename);
+	$jsoncount = 0;
+}
 
 ///////////////////////////////////////////////////
 $allfiles = scandir($datafolder);
@@ -126,40 +141,6 @@ foreach ($thefiles as $thisfile) {
 	$maxCell = $sheetObj->getHighestRowAndColumn();
 	$sheetData = $sheetObj->rangeToArray('A1:' . $maxCell['column'] . $maxCell['row'],null,true,true,false);
 
-/* forget about this now that we're looking for keywords
- 
-	//get rid of empty cells (part 2)
-	foreach($sheetData as $key => &$row) {
-		$row = array_filter($row,
-							function($cell) {
-								if ( !is_null($cell) && ($cell != "") ) return true;
-									else return false;
-							}
-			   );
-		if (count($row) == 0) {
-			unset($sheetData[$key]);
-		}
-	}
-	unset($row);
-*/
-
-	
-/* moved to function
-	$search = "Salaries and Benefits";
-//	$search = "alarie";
-	foreach($sheetData as $colnum => $row) {
-		foreach($row as $rownum => $thiscell) {
-			if ( strcasecmp($thiscell,$search) == 0 ) {
-				print "[".$colnum.",".$rownum."] ".$thiscell.PHP_EOL;
-								
-				break(2);
-			} else {
-//				print "NO MATCH (".$colnum.",".$rownum."): ".$thiscell.PHP_EOL;
-			}
-		}
-	}
-*/
-
 	$salaryrow = findRow("Salaries and Benefits");
 	$materialsrow = findRow("Materials and Supplies");
 	$equipmentrow = findRow("Equipment");
@@ -173,7 +154,6 @@ foreach ($thefiles as $thisfile) {
 	if (is_null($toreserverow)) { $toreserverow = findRow("Contributions to Reserves"); }
 	$toreservefundrow = findRow("Contributions to Reserve Funds");
 
-	
 	$interdivrecorow = findRow("Interdivisional Recoveries");
 	$provsubrow = findRow("Provincial Subsidies");
 	$fedsubrow = findRow("Federal Subsidies");
@@ -208,50 +188,115 @@ foreach ($thefiles as $thisfile) {
 //	print "approved positions in 2012 ".$sheetData[$approvedpositionsrow][$budget2012col].PHP_EOL;
 //	print "average salary in 2012 is $".number_format($sheetData[$salaryrow][$budget2012col]*1000/$sheetData[$approvedpositionsrow][$budget2012col]).PHP_EOL;
 
-	//output!
-	//FORMAT:	division , year , actual or projected or budget or outlook , revenue OR expense , type , amount
-
 	$division = str_replace('.xlsx','',$thisfile);
 	$division = str_replace('.xls','',$division);
 
-	foreach ($columnloop as $c) {
+	if ($outputaslist) {
+
+		//FORMAT:	division , year , actual or projected or budget or outlook , revenue OR expense , type , amount
+
+		foreach ($columnloop as $c) {
+			$year = substr($c,0,4);
+			$kind = substr($c,5);
+
+			//output revenues
+			$revarr = array_keys($revenues);
+			for ($i = 0; $i < count($revarr); $i++) {		
+				if (isset(${$revarr[$i]})) {
+					if (isset($sheetData[${$revarr[$i]}][${$columnvar[$c]}])) {
+							file_put_contents($outputrevenuesfilename,sprintf('"%s",%s,%s,%s,%s,%s'.PHP_EOL,$division,$year,$kind,"revenues",$revenues[$revarr[$i]], round($sheetData[${$revarr[$i]}][${$columnvar[$c]}]*1000,2) ),FILE_APPEND | LOCK_EX);
+					} else {
+	//print 'ZZ ('.$division.')('.$revarr[$i].')('.$c.') isnt set: '.$sheetData[${$revarr[$i]}][${$columnvar[$c]}].PHP_EOL;
+					}
+				}
+			}
+
+			//output expenses		
+			$exparr = array_keys($expenses);
+			for ($i = 0; $i < count($exparr); $i++) {		
+				if (isset(${$exparr[$i]})) {
+					if (isset($sheetData[${$exparr[$i]}][${$columnvar[$c]}])) {
+							file_put_contents($outputexpensesfilename,sprintf('"%s",%s,%s,%s,%s,%s'.PHP_EOL,$division,$year,$kind,"expenses",$expenses[$exparr[$i]], round($sheetData[${$exparr[$i]}][${$columnvar[$c]}]*1000,2) ),FILE_APPEND | LOCK_EX);
+					} else {
+	//print '** ('.$division.')('.$exparr[$i].')('.$c.') isnt set: '.$sheetData[${$exparr[$i]}][${$columnvar[$c]}].PHP_EOL;
+					}
+				}
+			}
+
+			//output positions
+			file_put_contents($outputpositionsfilename,sprintf('"%s",%s,%s,%s,%s'.PHP_EOL,$division,$year,$kind,"positions", $sheetData[$approvedpositionsrow][${$columnvar[$c]}]) ,FILE_APPEND | LOCK_EX);
+
+		}
+	} 
+	if ($outputaschart) {
+
+		//always 2012-budget
+		//FORMAT: division,salaries,materials,equipment,sales,interdiv,services,other,capital,ccrf,reserve,resfunds
+		//file_put_contents($outputchartfilename,"division,salaries,materials,equipment,sales,interdiv,services,other,capital,ccrf,reserve,resfunds".PHP_EOL,FILE_APPEND | LOCK_EX);
+		file_put_contents($outputchartfilename,sprintf('"%s",%s,%s,%s,%s,%s,%s,%s,%s,%s,%s,%s'.PHP_EOL,$division,
+			round($sheetData[$salaryrow][$budget2012col]*1000,2),
+			round($sheetData[$materialsrow][$budget2012col]*1000,2),
+			round($sheetData[$equipmentrow][$budget2012col]*1000,2),
+			round($sheetData[$salesrow][$budget2012col]*1000,2),
+			round($sheetData[$interdivchargesrow][$budget2012col]*1000,2),
+			round($sheetData[$servicesrow][$budget2012col]*1000,2),
+			round($sheetData[$otherrow][$budget2012col]*1000,2),
+			round($sheetData[$capitalrow][$budget2012col]*1000,2),
+			round($sheetData[$ccrfrow][$budget2012col]*1000,2),
+			round($sheetData[$toreserverow][$budget2012col]*1000,2),
+			round($sheetData[$toreservefundrow][$budget2012col]*1000,2)),
+			FILE_APPEND | LOCK_EX);
+	}
+
+	if ($outputasjson) {
+
+		unset($divisionarray);
+
+		$c = "2012 Budget";
+
 		$year = substr($c,0,4);
 		$kind = substr($c,5);
 
-		//output revenues
-		$revarr = array_keys($revenues);
-		for ($i = 0; $i < count($revarr); $i++) {		
-			if (isset(${$revarr[$i]})) {
-				if (isset($sheetData[${$revarr[$i]}][${$columnvar[$c]}])) {
-						file_put_contents($outputrevenuesfilename,sprintf('"%s",%s,%s,%s,%s,%s'.PHP_EOL,$division,$year,$kind,"revenues",$revenues[$revarr[$i]], round($sheetData[${$revarr[$i]}][${$columnvar[$c]}]*1000,2) ),FILE_APPEND | LOCK_EX);
-				} else {
-//print 'ZZ ('.$division.')('.$revarr[$i].')('.$c.') isnt set: '.$sheetData[${$revarr[$i]}][${$columnvar[$c]}].PHP_EOL;
-				}
-			}
-		}
-
-		//output expenses		
+		//loop through expense types
 		$exparr = array_keys($expenses);
-		for ($i = 0; $i < count($exparr); $i++) {		
+		$divisioncount = 0;
+		for ($i = 0; $i < count($exparr); $i++) {
 			if (isset(${$exparr[$i]})) {
-				if (isset($sheetData[${$exparr[$i]}][${$columnvar[$c]}])) {
-						file_put_contents($outputexpensesfilename,sprintf('"%s",%s,%s,%s,%s,%s'.PHP_EOL,$division,$year,$kind,"expenses",$expenses[$exparr[$i]], round($sheetData[${$exparr[$i]}][${$columnvar[$c]}]*1000,2) ),FILE_APPEND | LOCK_EX);
+				if ( isset($sheetData[${$exparr[$i]}][${$columnvar[$c]}]) && ($sheetData[${$exparr[$i]}][${$columnvar[$c]}] > 0) )  {
+					$divisionarray[$divisioncount]["name"] = $expenses[$exparr[$i]];
+					$divisionarray[$divisioncount]["size"] = round($sheetData[${$exparr[$i]}][${$columnvar[$c]}]*1000,2);
+//					$divisionarray[$divisioncount]["size"] = "10";
+					$divisioncount++;
 				} else {
 //print '** ('.$division.')('.$exparr[$i].')('.$c.') isnt set: '.$sheetData[${$exparr[$i]}][${$columnvar[$c]}].PHP_EOL;
 				}
+			} else {
+print "something else (".${$exparr[$i]}.")".PHP_EOL;
 			}
 		}
+		//var_dump($divisionarray);
 
-		//output positions
-		file_put_contents($outputpositionsfilename,sprintf('"%s",%s,%s,%s,%s'.PHP_EOL,$division,$year,$kind,"positions", $sheetData[$approvedpositionsrow][${$columnvar[$c]}]) ,FILE_APPEND | LOCK_EX);
-
+		$budgetarray[$jsoncount]["name"] = $division;
+		$budgetarray[$jsoncount]["children"] = $divisionarray;
+		$jsoncount++;
+print $jsoncount.PHP_EOL;
 	}
-	
+
 
 	unset($sheetData);
 	unset($sheetObj);
 	unset($objReader);
 	unset($objPHPExcel);
+}
+
+if ($outputasjson) {
+	$object = new stdClass();
+	$object->name = "2012 budget";
+	$object->children = $budgetarray;
+
+	//var_dump($object);
+	//print json_encode($object);
+	file_put_contents($outputjsonfilename,json_encode($object));
 }
 
 function findRow($search) {
